+++ Mods/Project Eternity Sorcerer Street/Terrains.cs
@@ -4,7 +4,7 @@ namespace ProjectEternity.GameScreens.SorcererStreetScreen
 {
     class ElementalTerrain : TerrainSorcererStreet
     {
-        public ElementalTerrain(int XPos, int YPos, int LayerIndex, float LayerDepth, int TerrainTypeIndex)
+        public ElementalTerrain(int XPos, int YPos, int LayerIndex, float LayerDepth, byte TerrainTypeIndex)
             : base(XPos, YPos, LayerIndex, LayerDepth, TerrainTypeIndex)
         {
 
@@ -29,7 +29,7 @@ namespace ProjectEternity.GameScreens.SorcererStreetScreen
 
     class GateTerrain : TerrainSorcererStreet
     {
-        public GateTerrain(int XPos, int YPos, int LayerIndex, float LayerDepth, int TerrainTypeIndex)
+        public GateTerrain(int XPos, int YPos, int LayerIndex, float LayerDepth, byte TerrainTypeIndex)
             : base(XPos, YPos, LayerIndex, LayerDepth, TerrainTypeIndex)
         {
